refactor: extract domain folder
Currently the types folder was used to contain domain types, with their associated functions, as well as actual type definitions.

This change extracts domain modules into a separate directory. The goal is to only have `.d.ts` files inside the types folder.
