+++ src/domain/package-manifest.ts
