+++ src/domain/project-manifest.ts
