+++ src/domain/scoped-registry.ts
