+++ src/io/upm-config-io.ts
@@ -3,8 +3,8 @@ import TOML from "@iarna/toml";
 import log from "../logger";
 import isWsl from "is-wsl";
 import execute from "../utils/process";
-import { addAuth, UpmAuth, UPMConfig } from "../types/upm-config";
-import { RegistryUrl } from "../types/registry-url";
+import { addAuth, UpmAuth, UPMConfig } from "../domain/upm-config";
+import { RegistryUrl } from "../domain/registry-url";
 import { CustomError } from "ts-custom-error";
 import { IOError } from "../common-errors";
 import { AsyncResult, Result } from "ts-results-es";
