+++ test/base64.test.ts
@@ -1,4 +1,4 @@
-import { decodeBase64, encodeBase64 } from "../src/types/base64";
+import { decodeBase64, encodeBase64 } from "../src/domain/base64";
 import fc from "fast-check";
 
 describe("base 64", () => {
