+++ test/jest.d.ts
@@ -1,7 +1,7 @@
 import { Stream } from "./mock-console";
-import { DomainName } from "../src/types/domain-name";
-import { SemanticVersion } from "../src/types/semantic-version";
-import { PackageUrl } from "../src/types/package-url";
+import { DomainName } from "../src/domain/domain-name";
+import { SemanticVersion } from "../src/domain/semantic-version";
+import { PackageUrl } from "../src/domain/package-url";
 
 declare global {
   namespace jest {
